+++ Characters/src/main/kotlin/com/fablesfantasyrp/plugin/characters/data/entity/EntityCharacterRepository.kt
@@ -2,5 +2,5 @@ package com.fablesfantasyrp.plugin.characters.data.entity
 
 import com.fablesfantasyrp.plugin.database.entity.EntityRepository
 
-interface EntityCharacterRepository : EntityRepository<ULong, Character>, CharacterRepository {
+interface EntityCharacterRepository : EntityRepository<Int, Character>, CharacterRepository {
 }
